app/requirements.txt
Babel==2.8.0
certifi==2019.11.28
Django==2.2.8
django-cors-headers==3.2.0
django-cors-middleware==1.5.0
django-filter==2.2.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
phonenumbers==8.11.1
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
